+++ TEST/ContainerExtensions/Config/ConfigFromAssembly.cs
@@ -11,7 +11,6 @@ using NUnit.Framework;
 
 namespace Solti.Utils.DI.Container.Setup.Tests
 {
-    using Annotations;
     using Internals;
 
     [TestFixture]