+++ TEST/ServiceProvider/GetService.cs
@@ -9,8 +9,6 @@ using NUnit.Framework;
 
 namespace Solti.Utils.DI.Injector.Tests
 {
-    using Annotations;
-
     public partial class InjectorTestsBase<TContainer>
     {
         [Test]