+++ history.md
@@ -102,6 +102,7 @@
   - *breaking:* new `IComposite` layout
   - *breaking:* `AbstractServiceEntry` moved to the public API
   - *breaking:* `ServiceReference` moved to the public API
+  - *breaking:* removed `Solti.Utils.DI.Annotations` namespace
   - *added:* `IAsyncDisposable` support
   - *done:* using nullable reference types
   - *fixed:* named services could not be queried via `ServiceProvider` ([details](https://github.com/Sholtee/injector/commit/bd97a0caf2763b3a28d4c511b2ee02006b072b3b ))